Add experimental feature for pooled pages
This commit adds the environment variable
RUBY_GC_EXPERIMENTAL_FEATURE_SIZE_POOL_POOLED_PAGES that turns on a new
pooled pages algorithm for incremental marking. When this environment
is not set, the old algorithm is used. When this environment is set,
the new algorithm is used.

The old algorithm is too optimistic, it assumes that all size pools will
be used at the same speed, so it marks very litle at each incremental
marking step. However, if if one size pool is filled, then there will be
a long pause in incremental marking. The new algorithm is pessimistic,
meaning it assumes that one size pool will be filled before the other
size pools are filled, so it will mark more object at each incremental
marking step.